+++ kp.go
@@ -508,102 +508,3 @@ func redact(s string, redactStrings []string) string {
 func noredact(s string, redactStrings []string) string {
 	return s
 }
-
-type SortByOpts func(s *string)
-
-// sort related funcs
-func GetKeySortStr(opts ...SortByOpts) *string {
-	sortStr := ""
-	for _, opt := range opts {
-		opt(&sortStr)
-	}
-	return &sortStr
-}
-
-func buildSortOpts(val string) SortByOpts {
-	return func(s *string) {
-		*s += "," + val
-		// remove the extra comma appended in the begining of the string
-		*s = strings.TrimLeft(*s, ",")
-	}
-}
-
-// sort by id
-func WithID() SortByOpts {
-	return buildSortOpts("id")
-}
-func WithIDDesc() SortByOpts {
-	return buildSortOpts("-id")
-}
-
-// sort by creation date
-func WithCreationDate() SortByOpts {
-	return buildSortOpts("creationDate")
-}
-
-func WithCreationDateDesc() SortByOpts {
-	return buildSortOpts("-creationDate")
-}
-
-// sort by deletionDate
-func WithDeletionDate() SortByOpts {
-	return buildSortOpts("deletionDate")
-}
-
-func WithDeletionDateDesc() SortByOpts {
-	return buildSortOpts("-deletionDate")
-}
-
-// sort by expirationDate
-func WithExpirationDate() SortByOpts {
-	return buildSortOpts("expirationDate")
-}
-
-func WithExpirationDateDesc() SortByOpts {
-	return buildSortOpts("-expirationDate")
-}
-
-// sort by extractable
-func WithExtractable() SortByOpts {
-	return buildSortOpts("extractable")
-}
-
-func WithExtractableDesc() SortByOpts {
-	return buildSortOpts("-extractable")
-}
-
-// sort by imported
-func WithImported() SortByOpts {
-	return buildSortOpts("imported")
-}
-
-func WithImportedDesc() SortByOpts {
-	return buildSortOpts("-imported")
-}
-
-// sort by lastRotateDate
-func WithLastRotateDate() SortByOpts {
-	return buildSortOpts("lastRotateDate")
-}
-
-func WithLastRotateDateDesc() SortByOpts {
-	return buildSortOpts("-lastRotateDate")
-}
-
-// sort by lastUpdateDate
-func WithLastUpdateDate() SortByOpts {
-	return buildSortOpts("lastUpdateDate")
-}
-
-func WithLastUpdateDateDesc() SortByOpts {
-	return buildSortOpts("-lastUpdateDate")
-}
-
-// sort by state
-func WithState() SortByOpts {
-	return buildSortOpts("state")
-}
-
-func WithStateDesc() SortByOpts {
-	return buildSortOpts("-state")
-}
